修改bug

+++ centit-utils/src/main/java/com/centit/support/common/CachedMap.java
@@ -71,7 +71,7 @@ public class CachedMap<K, T> extends AbstractCachedObject<Map<K, T>> {
         this.refresher = refresher;
     }
 
-    public void setFreshPeriod(int freshPeriod) {
+    public void setFreshPeriod(long freshPeriod) {
         this.freshPeriod = freshPeriod;
     }
