+++ centit-utils/src/main/java/com/centit/support/common/CachedObject.java
@@ -28,7 +28,7 @@ public class CachedObject<T> extends AbstractCachedObject<T> {
      * @param refresher   重新获取代码的接口
      * @param freshPeriod 保鲜时间，单位为秒
      */
-    public CachedObject(Supplier<T> refresher, int freshPeriod) {
+    public CachedObject(Supplier<T> refresher, long freshPeriod) {
         this.target = null;
         this.evicted = true;
         this.refresher = refresher;
@@ -57,7 +57,7 @@ public class CachedObject<T> extends AbstractCachedObject<T> {
     /**
      * @param freshPeriod 刷新周期 单位秒
      */
-    public void setFreshPeriod(int freshPeriod) {
+    public void setFreshPeriod(long freshPeriod) {
         this.freshPeriod = freshPeriod;
     }
